Change unit tests and files structure
